Source code submitted by Gowth from Feb. 16, 2011
Feb. 9th 2011     1. In some of the schedules, the last transition does
not have any code executed in it     * this leaves a blank line in the
transition info     * it causes a small bug in the program where the
last transition info shows the Final Error instead of showing a blank
line     * Possible solutions:         i. Remove the transition
completely when we see nothing is executed (this will our program
inconsistent with the test report            produced)         ii. OR we
can add a special case where we skip over these transitions when trying
to highlight the piece of code in the source code             when we
click that transition  2. Created a menu option called "Align" and
Finished the option to Align by both     * In this menu item you can
choose to align the top panel         i. Top - This is the original way
we aligned the schedules         ii. Bottom- This is aligning the
schedules from the bottom up (the errors are all aligned together)
Feb 10th 2011     1. New sorting mechanism needed for Sort by Bug Type
(is Now Called "Similarity: Dependent")       * Jeremy and I have
discussed a Sort where it first sorts by Bug Type and then after sorts
this list by similar transitions that lead to that bug type.     Feb
14th 2011     1. Current Similarity Code: Similarity Code.txt     Feb
15th 2011     1. When a transition does not execute any code we then
show its info as "JPF has no sources for bytecode executed (Most Likely
from an External Library)" --> Completed  2. When "Highlight Other Code"
option is enabled we no longer show other transitions with the same code
executed with a gold highlight.  Instead we just brighten the original
thread color so  we do not loose information on what thread executed
that code. ---> Completed  3. In the error info transition where it
shows which threads caused the error we did not highlight the error line
with the corresponding thread color.  This has been changed so the
correct thread  color is used to highlight these lines.  ---> Completed
Feb 16th 2011     1. A New Sort has been added called "Similarity:
Dependent" --> Completed       * This sort uses schedule 0 as the
reference points and then groups similar schedules based on a new
similarity metric